Warning System implementiert und Level System vorbereitet
Das Warning System ist soweit fertig, jedoch müssen noch tempban, tempmute usw. implementiert werden, sodass auch autopunishement funktioniert. Ebenso muss noch ein Command implementiert werden, der das definieren von autopunishements erlaubt.
Zum Thema Levelsystem: Die Klasse ist schon da, jedoch müssen noch Vergabe von XP, Aufrufen der Klasse und Ändern des Levels je nach XP entwickelt werden. Außerdem sollte man sich Gedanken darüber machen, wie man später die Levelcard designen möchte!

+++ src/main/java/commands/moderation/Clear.java
@@ -3,18 +3,44 @@ package commands.moderation;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
 
+import commands.Command;
 import components.AnswerEngine;
+import components.Developers;
+import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.entities.Guild;
 import net.dv8tion.jda.api.entities.Member;
 import net.dv8tion.jda.api.entities.Message;
 import net.dv8tion.jda.api.entities.TextChannel;
+import net.dv8tion.jda.api.events.interaction.SlashCommandEvent;
+import net.dv8tion.jda.api.interactions.commands.OptionType;
+import net.dv8tion.jda.api.interactions.commands.build.CommandData;
+import net.dv8tion.jda.api.interactions.commands.build.OptionData;
 
-public class Clear {
+public class Clear implements Command{
 
-	public Clear(Guild guild, Member member, TextChannel channel, String argument) {
-		List<Message> messages = channel.getHistory().retrievePast(Integer.parseInt(argument)+1).complete();
+	@Override
+	public void perform(SlashCommandEvent event) {
+		final Member member = event.getMember();
+		if (!member.hasPermission(Permission.MESSAGE_MANAGE) && !Developers.getInstance().developers.contains(event.getMember().getId())) {
+			event.replyEmbeds(AnswerEngine.getInstance().fetchMessage("/commands/moderation/clear:nopermission")).queue();
+			return;
+		}
+		final TextChannel channel = event.getTextChannel();
+		final int count = Integer.parseInt(event.getOption("count").getAsString());
+		List<Message> messages = channel.getHistory().retrievePast(count).complete();
 		channel.deleteMessages(messages).queue();
-		channel.sendMessageEmbeds(AnswerEngine.getInstance().fetchMessage("/commands/moderation/clear:done", guild, member)).queue(response -> response.delete().queueAfter(3, TimeUnit.SECONDS));
+		event.replyEmbeds(AnswerEngine.getInstance().fetchMessage("/commands/moderation/clear:done")).queue(response -> response.deleteOriginal().queueAfter(3, TimeUnit.SECONDS));
+	}
+
+	@Override
+	public CommandData initialize(Guild guild) {
+		CommandData command = new CommandData("clear", "Clears a specific number of messages from this channel!").addOptions(new OptionData(OptionType.INTEGER, "count", "Hand over the number of messages you want to delete!"));
+		return command;
+	}
+
+	@Override
+	public String getHelp() {
+		return "Use this command to clear messages from the channel you use the command in.";
 	}
 
 }
